requirements.txt
pandas
matplotlib
seaborn
openpyxl
jupyter
great_expectations
scikit-learn
plotly
nbformat
statsmodels
prophet
torch
pytorch-lightning
xgboost
shap
pydantic>=2.11.0
mlflow>=3.11.1
optuna
evidently>=0.7.21
apache-airflow>=3.2.1